Nearly ready to try out our bulk insert command.

+++ Startitecture.Orm.Testing.Schema/dbo/Tables/IntegerElement.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[IntegerElement] (
     [IntegerElementId] BIGINT NOT NULL,
-    [Value]            INT    NOT NULL,
+    [Value]            BIGINT    NOT NULL,
     CONSTRAINT [PK_IntegerElement] PRIMARY KEY CLUSTERED ([IntegerElementId] ASC),
     CONSTRAINT [FK_IntegerElement_FieldValueElement] FOREIGN KEY ([IntegerElementId]) REFERENCES [dbo].[FieldValueElement] ([FieldValueElementId])
 );
